refactor: remove application settings shim
