+++ tests/unit/application/test_feed_polling_service.py
@@ -8,7 +8,7 @@ import pytest
 from astrbot_plugin_rsshub.src.application.services.feed_polling_service import (
     FeedPollingService,
 )
-from astrbot_plugin_rsshub.src.application.settings import RSSSettings
+from astrbot_plugin_rsshub.src.shared.settings import RSSSettings
 from astrbot_plugin_rsshub.src.domain.entities.feed import Feed
 from astrbot_plugin_rsshub.src.infrastructure.fetcher.rss.parser import EntryParsed
