+++ tests/unit/infrastructure/test_route_knowledge_repository.py
@@ -7,7 +7,7 @@ import pytest
 from astrbot_plugin_rsshub.src.application.ports.route_knowledge import (
     RouteKnowledgeDocument,
 )
-from astrbot_plugin_rsshub.src.application.settings import RouteKnowledgeSettings
+from astrbot_plugin_rsshub.src.shared.settings import RouteKnowledgeSettings
 from astrbot_plugin_rsshub.src.infrastructure.knowledge.astrbot_kb_repository import (
     AstrBotRouteKnowledgeRepository,
 )
